pyproject.toml
[tool.black]
line-length = 100
target-version = ['py39']


[tool.poetry]
name = "vane"
version = "0.9"
description = "Network Certification Tool"
authors = ["Arista Networks, Inc."]
readme = "README.md"
classifiers=["Programming Language :: Python :: 3", "Operating System :: OS Independent",]
include=["vane"]
keywords=["vane", "pytest"]


[tool.poetry.dependencies]
python = "^3.9"
cvprac = "^1.3.0"
jinja2 = "^3.1.2"
mdutils = "^1.5.1"
netmiko = "^4.1.2"
pyeapi = "^0.8.4"
python-docx = "^0.8.11"
pyyaml = "^6.0"
coverage = "^7.2.2"
coverage-badge = "^1.1.0"
mypy = "^1.1.1"
pytest = "^7.2.2"
pytest-cache = "^1.0"
pytest-cov = "^4.0.0"
pytest-excel = "^1.5.0"
pytest-html = "^3.2.0"
pytest-json = "^0.4.0"
pytest-xdist = "^3.2.1"
pytest-mock = "^3.10.0"

[tool.poetry.dev-dependencies]
pylint = "2.17.0"
flake8 = "6.0.0"
black = "23.1.0"


[tool.poetry.scripts]
vane = "vane.vane_cli:main"
